Initial commit
Added the main.py file. The 'main.py' file is the main script file required for the proper functioning of the application. The file contains the entire code base of the application. This is the only version that is ever posted for this application / project. It is said so because the application is meant to be a high school project, so no more updates or adding of features by the author. But, contribution by you is open, you can add new features to it or even make it better. Just after the update, create a pull request, or send the author the updated version of the code at email : rdofficial192@gmail.com